+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - ResourceName parameter of Add-LabMachineDefinition now actually supported. Fixes #23
   - No interaction is done in AL using the resource name. This is only for the purpose of
     deploying the same lab on the same host with different resource names (VM names, switch names)
+- Enabling configuration of allowed inbound IP addresses for Azure load balancer
   
 ### Bug Fixes
 - Get-LabInternetFile did not work on Azure when the Uri did not contain a file name like 'https://go.microsoft.com/fwlink/?Linkid=85215'.